Refactor and improve int editor
